refactor(library): rename specification module to index
shorter import paths in other modules
